Microproyecto2/requirements.txt
flask==2.3.3
gunicorn==21.2.0
Pillow==10.3.0
numpy==1.23.5
mxnet==1.9.1
gluoncv==0.10.5.post0
